Fix level select, planet size

+++ rund-um-die-welten/planet/planet_v_2.gd
@@ -15,6 +15,7 @@ extends Area2D
 
 @export var clockwise = true
 
+@export var collision_enabled = true
 func _ready():
 	var shape = $CollisionShape2D.shape
 	# Sprite skalieren falls notwendig?
@@ -27,19 +28,19 @@ func _ready():
 	# change size of collision area and animation based on which planet is chosen
 	match planet_type:
 		GlobalVariables.planets.DeathStar:
-			size_scale = 1.0
+			size_scale = 1.5
 			$death_star.play("default")
 			$death_star.visible = true
 		GlobalVariables.planets.Mars:
-			size_scale = 1.0
+			size_scale = 1.5
 			$mars.play("default")
 			$mars.visible = true
 		GlobalVariables.planets.Earth:
-			size_scale = 2.8
+			size_scale = 4.2
 			$earth.play("default")
 			$earth.visible = true
 		GlobalVariables.planets.Sun:
-			size_scale = 3.2
+			size_scale = 4.8
 			$sun.play("default")
 			$sun.visible = true
 	
@@ -54,14 +55,15 @@ func _input_event(viewport, event, shape_idx):
 
 
 func _on_area_entered(area: Area2D) -> void:
-	# if area in player group entered, kill player
-	if area.is_in_group("player"):
-		if win_planet:
-			GlobalVariables.emit_level_done(win_message)
-		else:
-			if death_message_overwrite:
-				GlobalVariables.emit_player_died(death_message_overwrite)
+	if collision_enabled:
+		# if area in player group entered, kill player
+		if area.is_in_group("player"):
+			if win_planet:
+				GlobalVariables.emit_level_done(win_message)
 			else:
-				GlobalVariables.emit_player_died("You crashed into the " + GlobalVariables.planet_names[planet_type])
-			print(area.position)
+				if death_message_overwrite:
+					GlobalVariables.emit_player_died(death_message_overwrite)
+				else:
+					GlobalVariables.emit_player_died("You crashed into the " + GlobalVariables.planet_names[planet_type])
+				print(area.position)
